[rename] handleApiError -> surfGuard
