+++ src/app/api/gatherings/participants/route.ts
@@ -1,7 +1,7 @@
 import { NextRequest, NextResponse } from 'next/server';
 import { externalClient } from '@/lib/api/clientFetchers';
 import { EXTERNAL_PATHS } from '@/lib/api/apiPaths';
-import { handleApiError } from '@/lib/api/handleApiError';
+import { handleApiError } from '@/lib/api/surfGuard';
 
 /**
  * 모임 참여자 목록 조회
