+++ src/app/api/reviews/route.ts
@@ -1,6 +1,6 @@
 import { NextRequest, NextResponse } from 'next/server';
 import { EXTERNAL_PATHS } from '@/lib/api/apiPaths';
-import { handleApiError } from '@/lib/api/handleApiError';
+import { handleApiError } from '@/lib/api/surfGuard';
 import { externalClient } from '@/lib/api/clientFetchers';
 
 /**
